+++ test/it/pure-svelte.test.ts
